Stamp every update packet with the map the client is rendering
A GM standing on a deck who typed ".npc add 6086" killed his own client
outright: ERROR #132, ACCESS_VIOLATION reading [NULL+0x1A0].

Object::SendCreateUpdateToPlayer was the one create path never converted to
Player::GetClientMapId(). It stamped the packet with player->GetMapId(), which
for a passenger is the VESSEL'S map -- an id his client has never been told,
has no terrain for, and cannot look up. It is also the path Map::Add takes to
show a newly spawned object to someone already standing there, via the singular
sweep in Player::UpdateVisibilityOf.

Proven on the wire rather than deduced. In the whole capture of the crashing
session every SMSG_UPDATE_OBJECT carries map 0 -- including the vessel's crew,
announced to him seconds earlier after he walked aboard -- except one:

    48 02 | 01 00 00 00 | 02 | F7 8B3A82 C617 30F1
    ^^^^^   blockCount   CREATE  guid F13017C600823A8B (entry 6086)
    map 0x0248 = 584 "Transport: Menethil to Theramore"

and the client answered that packet with its shutdown triple
(CMSG_DISCARDED_TIME_SYNC_ACKS, two CMSG_CANCEL_TRADE) and no logout request.

Two more sites carried the same defect and would have fired on any passenger:

- Player::UpdateForQuestWorldObjects, on a quest-object refresh aboard.
- Player::FlushPendingEmoteRefresh, which used the EMITTER's map id for a
  packet addressed to the receiver.

Object::BuildUpdateDataForPlayer was already correct and is only re-cast.

Every UpdateData construction now goes through GetClientMapId() and an explicit
static_cast<uint16>: the constructor takes uint16 and every caller was handing
it a uint32, which MSVC reports as C4244.

+++ src/game/WorldHandlers/GridNotifiers.h
@@ -46,7 +46,7 @@ namespace MaNGOS
         GuidSet i_clientGUIDs;
         std::set<WorldObject*> i_visibleNow;
 
-        explicit VisibleNotifier(Camera &c) : i_camera(c), i_clientGUIDs(c.GetOwner()->m_clientGUIDs), i_data(c.GetOwner()->GetClientMapId()) {}
+        explicit VisibleNotifier(Camera &c) : i_camera(c), i_clientGUIDs(c.GetOwner()->m_clientGUIDs), i_data(static_cast<uint16>(c.GetOwner()->GetClientMapId())) {}
         template<class T> void Visit(GridRefManager<T>& m);
         void Visit(CameraMapType& /*m*/) {}
         void Notify(void);
